grafana/requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
paho-mqtt==2.0.0
pillow==10.2.0
ping3==4.0.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
